Moved test helpers to __testUtils__ folder. Added stringifyEntityRef as a dependency in MembersListCard
